Fixed casted-lvalue warnings/errors for GCC 4+
The latest versions of GCC do not like casted-lvalues assignments like:

((int)i) = 5;

To fix this, the WAM engine PCell union was extended to include backward-compatibility types. This eliminated the need for casting in many cases. In a few additional places, castes were removed or shifted, or explicit lvalue macros were defined.

+++ core/alsp_src/generic/bsio.c
@@ -2346,7 +2346,7 @@ sio_socket_open()
 	    SIO_TYPE(buf) = SIO_TYPE_SOCKET_DGRAM;
 	    if (isserver && domain == AF_INET) {
 		struct sockaddr_in *sa;
-		SIO_SOCKET_ADDRESS(buf) = malloc(sizeof (struct sockaddr_in));
+		SIO_SOCKET_ADDRESS(buf) = (long) malloc(sizeof (struct sockaddr_in));
 		SIO_SOCKET_ADDRESS_LEN(buf) = sizeof (struct sockaddr_in);
 		sa = (struct sockaddr_in *) SIO_SOCKET_ADDRESS(buf);
 		memset(sa, 0, sizeof (struct sockaddr_in *));
